refactor api routes to app route handlers
